Django/requirements.txt
asgiref==3.4.1
attrs==21.4.0
autobahn==22.6.1
Automat==20.2.0
Brotli==1.0.9
certifi==2022.6.15
cffi==1.15.1
channels==2.4.0
ciso8601==2.2.0
click==8.1.3
colorama==0.4.5
constantly==15.1.0
cryptography==37.0.4
daphne==2.5.0
dash==2.6.0
dash-bootstrap-components==0.13.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Django==3.2.14
django-plotly-dash==2.0.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
dpd-components==0.1.0
Flask==2.1.3
Flask-Compress==1.12
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.12.0
incremental==21.3.0
influxdb-client==1.30.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
opcua==0.98.13
paho-mqtt==1.6.1
plotly==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
python-dateutil==2.8.2
pytz==2022.1
Rx==3.2.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
tenacity==8.0.1
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.1.1
urllib3==1.26.10
Werkzeug==2.0.3
zipp==3.8.1
zope.interface==5.4.0
